change int to float

+++ homework2.py
@@ -1,3 +1,3 @@
 Celsius = input("Enter Celsius Temperature : ")
-Fahrenheit = int(Celsius) * 9/5 + 32
+Fahrenheit = float(Celsius) * 9/5 + 32
 print("Fahrenheit Temperature : ", Fahrenheit)
